+++ src/bika/lims/interfaces/__init__.py
@@ -42,16 +42,6 @@ class ISenaiteSite(IBikaLIMS):
     """
 
 
-class IAutoGenerateID(Interface):
-    """Auto-generate ID with ID server
-    """
-
-
-class IMultiCatalogBehavior(Interface):
-    """Support multiple catalogs for Dexterity contents
-    """
-
-
 class IActionHandlerPool(Interface):
     """Marker interface for the ActionHandlerPool utility
     """
@@ -74,11 +64,6 @@ class IDoNotSupportSnapshots(Interface):
 
 class IAuditLogCatalog(Interface):
     """Audit Log Catalog
-    """
-
-
-class IGenerateID(Interface):
-    """Marker Interface to generate an ID
     """
 
 
@@ -581,35 +566,6 @@ class IBikaCatalogWorksheetListing(Interface):
 class IBikaCatalogReport(Interface):
     """Marker interface for bika_catalog_report
     """
-
-
-class IIdServer(Interface):
-    """Marker Interface for ID server
-    """
-
-    def generate_id(self, portal_type, batch_size=None):
-        """Generate a new id for 'portal_type'
-        """
-
-
-class IIdServerVariables(Interface):
-    """Marker interfaces for variables generator for ID Server
-    """
-
-    def get_variables(self, **kw):
-        """Returns a dict with variables
-        """
-
-
-class IIdServerTypeID(Interface):
-    """Marker interface for type id resolution for ID Server
-    """
-
-    def get_type_id(self, **kw):
-        """Returns the type id for the context passed in the constructor, that
-        is used for custom ID formatting, regardless of the real portal type of
-        the context. Return None if no type id can be resolved by this adapter
-        """
 
 
 class IDisplayListVocabulary(Interface):
