--- src/bika/lims/profiles/default/componentregistry.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0"?>
-<componentregistry>
-  <utilities>
-
-    <!-- remove after 2.0.0rc3 -->
-    <utility
-      remove="True"
-      component="bika.lims.controlpanel.bika_idserver.bika_idserver"
-      interface="bika.lims.interfaces.IIdServer"
-      provides="bika.lims.interfaces.IIdServer"/>
-
-  </utilities>
-</componentregistry>
