--- src/bika/lims/subscribers/objectadded.py
@@ -1,29 +1,0 @@
-# -*- coding: utf-8 -*-
-#
-# This file is part of SENAITE.CORE.
-#
-# SENAITE.CORE is free software: you can redistribute it and/or modify it under
-# the terms of the GNU General Public License as published by the Free Software
-# Foundation, version 2.
-#
-# This program is distributed in the hope that it will be useful, but WITHOUT
-# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
-# FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
-# details.
-#
-# You should have received a copy of the GNU General Public License along with
-# this program; if not, write to the Free Software Foundation, Inc., 51
-# Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
-#
-# Copyright 2018-2021 by it's authors.
-# Some rights reserved, see README and LICENSE.
-
-from bika.lims.idserver import renameAfterCreation
-from bika.lims import logger
-
-
-def auto_generate_id(obj, event):
-    """Generate ID with the IDServer from senaite.core
-    """
-    logger.info("Auto-Generate ID for {}".format(repr(obj)))
-    renameAfterCreation(obj)
