+++ src/senaite/core/idserver/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+from .idserver import *  # noqa
